# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VendlySys - SaaS multiempresa para salões e clínicas
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login do usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login bem-sucedido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Credenciais inválidas
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Dados do usuário logado
      responses:
        "200":
          description: Usuário atual
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsuarioComEmpresa"
        "401":
          description: Não autenticado
...
  /agendamentos:
    get:
      operationId: listAgendamentos
      tags: [agendamentos]
      summary: Lista agendamentos da empresa
      parameters:
        - name: data_inicio
          in: query
          schema:
            type: string
            format: date
        - name: data_fim
          in: query
          schema:
            type: string
            format: date
        - name: profissional_id
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [agendado, confirmado, concluido, cancelado]
      responses:
        "200":
          description: Lista de agendamentos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Agendamento"
    post:
      operationId: createAgendamento
      tags: [agendamentos]
      summary: Cria novo agendamento
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgendamentoBody"
      responses:
        "201":
          description: Agendamento criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agendamento"
  /agendamentos/{id}:
    get:
      operationId: getAgendamento
      tags: [agendamentos]
      summary: Busca agendamento por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Agendamento encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agendamento"
    put:
      operationId: updateAgendamento
      tags: [agendamentos]
      summary: Atualiza agendamento
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAgendamentoBody"
      responses:
        "200":
          description: Agendamento atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agendamento"
    delete:
      operationId: deleteAgendamento
      tags: [agendamentos]
      summary: Cancela agendamento
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Agendamento cancelado
...
components:
  schemas:
...
    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        senha:
          type: string
      required: [email, senha]
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        usuario:
          $ref: "#/components/schemas/UsuarioComEmpresa"
      required: [token, usuario]
    UsuarioComEmpresa:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        email:
          type: string
        perfil:
          type: string
          enum: [admin, agenda]
        empresa_id:
          type: integer
        empresa:
          $ref: "#/components/schemas/Empresa"
      required: [id, nome, email, perfil, empresa_id]
    Empresa:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        slug:
          type: string
        cor_primaria:
          type: string
        cor_secundaria:
          type: string
        logo_url:
          type: string
        telefone:
          type: string
        endereco:
          type: string
      required: [id, nome, slug]
...
    Agendamento:
      type: object
      properties:
        id:
          type: integer
        empresa_id:
          type: integer
        cliente_id:
          type: integer
        cliente_nome:
          type: string
        profissional_id:
          type: integer
        profissional_nome:
          type: string
        servico:
          type: string
        data:
          type: string
          format: date
        hora_inicio:
          type: string
        hora_fim:
          type: string
        valor:
          type: number
        status:
          type: string
          enum: [agendado, confirmado, concluido, cancelado]
        observacoes:
          type: string
        forma_pagamento:
          type: string
          enum: [dinheiro, pix, cartao_credito, cartao_debito, transferencia]
        pago:
          type: boolean
      required: [id, empresa_id, servico, data, hora_inicio, status]
    CreateAgendamentoBody:
      type: object
      properties:
        cliente_id:
          type: integer
        profissional_id:
          type: integer
        servico:
          type: string
        data:
          type: string
          format: date
        hora_inicio:
          type: string
        hora_fim:
          type: string
        valor:
          type: number
        forma_pagamento:
          type: string
          enum: [dinheiro, pix, cartao_credito, cartao_debito, transferencia]
        observacoes:
          type: string
      required: [servico, data, hora_inicio]
    UpdateAgendamentoBody:
      type: object
      properties:
        cliente_id:
          type: integer
        profissional_id:
          type: integer
        servico:
          type: string
        data:
          type: string
          format: date
        hora_inicio:
          type: string
        hora_fim:
          type: string
        valor:
          type: number
        status:
          type: string
          enum: [agendado, confirmado, concluido, cancelado]
        forma_pagamento:
          type: string
          enum: [dinheiro, pix, cartao_credito, cartao_debito, transferencia]
        pago:
          type: boolean
        observacoes:
          type: string
...